Fixed the serialization. Fixed the issues with the caching of the game data.

+++ src/Utility.h
@@ -18,7 +18,7 @@ struct Handle {
 struct Generation {
 	bool slot_Taken = false;
 	// Default generation 1
-	uint16_t generation = 1;
+	uint64_t generation = 1;
 };
 int count_Active_Handles(Generation generations[], int size);
 bool compare_Handles(Handle handle_1, Handle handle_2);
